Primite scope caching
