+++ src/react/index.ts
@@ -1,4 +1,7 @@
+
+export { MoleculeScopeOptions } from "../shared/MoleculeScopeOptions";
 export * from "./ScopeProvider";
 export * from "./useMolecule";
 export * from "./useScopes";
 export * from "./useStore";
+
